+++ src/cadenzaanalytics/__init__.py
@@ -12,6 +12,7 @@ from logging.config import dictConfig
 
 from cadenzaanalytics.cadenza_analytics_extension import CadenzaAnalyticsExtension
 from cadenzaanalytics.cadenza_analytics_extension_service import CadenzaAnalyticsExtensionService
+
 from cadenzaanalytics.data.analytics_extension import AnalyticsExtension
 from cadenzaanalytics.data.attribute_group import AttributeGroup
 from cadenzaanalytics.data.attribute_role import AttributeRole
@@ -22,11 +23,18 @@ from cadenzaanalytics.data.geometry_type import GeometryType
 from cadenzaanalytics.data.measure_aggregation import MeasureAggregation
 from cadenzaanalytics.data.parameter import Parameter
 from cadenzaanalytics.data.parameter_type import ParameterType
+
 from cadenzaanalytics.request.request_metadata import RequestMetadata
-from cadenzaanalytics.response.csv_response import CsvResponse
-from cadenzaanalytics.response.row_wise_mapping_csv_response import RowWiseMappingCsvResponse
-from cadenzaanalytics.response.error_response import ErrorResponse
+from cadenzaanalytics.request.view_parameter import ViewParameter
+
+from cadenzaanalytics.response.data_response import DataResponse
+from cadenzaanalytics.response.enrichment_response import EnrichmentResponse
+from cadenzaanalytics.response.missing_metadata_strategy import MissingMetadataStrategy
 from cadenzaanalytics.response.image_response import ImageResponse
+from cadenzaanalytics.response.text_response import TextResponse
+from cadenzaanalytics.response.url_response import UrlResponse
+from cadenzaanalytics.response.error_response import ErrorResponse
+
 from cadenzaanalytics.version import __version__