+++ src/cadenzaanalytics/data/data_type.py
@@ -1,5 +1,6 @@
 from enum import Enum
 
+from pandas import DataFrame
 
 # pylint: disable=duplicate-code
 class DataType(Enum):
@@ -14,7 +15,7 @@ class DataType(Enum):
     def __str__(self):
         return self.value
 
-    def pandas_type(self):
+    def pandas_type(self) -> str:
         """Return the corresponding pandas data type for the given value.
 
         Returns
@@ -27,3 +28,19 @@ class DataType(Enum):
         if self.value == 'float64':
             return "Float64"
         return "string"
+
+
+    @classmethod
+    def from_pandas_dtype(cls, dtype: DataFrame.dtypes):
+        """Return the cadenza analytics data type for given pandas data type.
+
+        Returns
+        -------
+        DataType
+            The cadenza analytics data type corresponding to the given value.
+        """
+        if dtype == 'int64':
+            return DataType.INT64
+        if dtype == 'float64':
+            return DataType.FLOAT64
+        return DataType.STRING