+++ src/cadenzaanalytics/data/extension_type.py
@@ -2,11 +2,11 @@ from enum import Enum
 
 
 class ExtensionType(Enum):
-    """A class representing the different types of extensions such as visualization, enrichment, calculation.
+    """A class representing the different types of extensions such as visual, enrichment, data.
     """
-    VISUALIZATION = "visualization"
+    VISUAL = "visualization"
     ENRICHMENT = "enrichment"
-    CALCULATION = "calculation"
+    DATA = "calculation"
 
     def __str__(self):
         return self.value