+++ src/cadenzaanalytics/response/enrichment_response.py
@@ -1,0 +1,44 @@
+from typing import List
+
+from pandas import DataFrame
+
+from cadenzaanalytics.data.column_metadata import ColumnMetadata
+from cadenzaanalytics.data.attribute_group import AttributeGroup
+from cadenzaanalytics.response.csv_response import CsvResponse
+from cadenzaanalytics.response.missing_metadata_strategy import MissingMetadataStrategy
+
+
+class EnrichmentResponse(CsvResponse):
+    """A class representing an enrichment response from an extension.
+
+    Parameters
+    ----------
+    CsvResponse : type
+        The data response type from which EnrichmentResponse inherits.
+    """
+    def __init__(self, data: DataFrame, column_metadata: List[ColumnMetadata], missing_metadata_strategy: MissingMetadataStrategy = MissingMetadataStrategy.ADD_DEFAULT_METADATA):
+        super().__init__(data, column_metadata, missing_metadata_strategy)
+
+    def get_response(self, original_column_metadata: List[ColumnMetadata], original_data: DataFrame):
+        """Get the enrichment response.
+
+        Returns
+        -------
+        Response
+            The enrichment response.
+        """
+        if self._is_runtime_validation_active:
+            self._validate_response()
+
+        return super().get_response(original_column_metadata, original_data)
+
+
+    def _validate_response(self):
+        has_id_column_defined = False
+
+        for column in self._column_meta_data:
+            if column.attribute_group_name == AttributeGroup.ID_ATTRIBUTE_GROUP_NAME:
+                has_id_column_defined = True
+
+        if not has_id_column_defined:
+            raise Exception(f"Identifier column missing in metadata definition. This is mandatory for an enrichment.")